Add IRC bot
As per the original concept for this project, support connecting to
to IRC, running gerrit-nag when requested and post the results.

Would have liked to use python3 for this but seems like the twisted
IRC module has python3 problems; see here:
https://twistedmatrix.com/trac/ticket/6320

Since it's python2 it's calling gerrit-nag.py with a system call,
which is not ideal, but good enough for now.

May in future extend this so it can do other useful things, at which
point it might be better separated into a different repo, but worry
about that later.